+++ src/main/java/de/erdbeerbaerlp/curseforgeBot/storage/json/Root.java
@@ -4,7 +4,7 @@ package de.erdbeerbaerlp.curseforgeBot.storage.json;
 import java.util.Arrays;
 
 public class Root {
-    public Integer[] projects = new Integer[0];
+    public Long[] projects = new Long[0];
     public Settings settings = new Settings();
 
     @Override
